# ui/desktop/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "daisy-server",
    "description": "An AI agent",
    "contact": {
      "name": "Block",
      "email": "ai-oss-tools@block.xyz"
    },
    "license": {
      "name": "Apache-2.0"
    },
    "version": "1.8.0"
  },
  "paths": {
...
    "/context/manage": {
      "post": {
        "tags": [
          "Context Management"
        ],
        "operationId": "manage_context",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContextManageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Context managed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContextManageResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing API key"
          },
          "412": {
            "description": "Precondition failed - Agent not available"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
    "/recipes/create": {
      "post": {
        "tags": [
          "Recipe Management"
        ],
        "summary": "Create a Recipe configuration from the current session",
        "operationId": "create_recipe",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRecipeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Recipe created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateRecipeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "412": {
            "description": "Precondition failed - Agent not available"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/recipes/decode": {
      "post": {
        "tags": [
          "Recipe Management"
        ],
        "operationId": "decode_recipe",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecodeRecipeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Recipe decoded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecodeRecipeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/recipes/delete": {
      "post": {
        "tags": [
          "Recipe Management"
        ],
        "operationId": "delete_recipe",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteRecipeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Recipe deleted successfully"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing API key"
          },
          "404": {
            "description": "Recipe not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/recipes/encode": {
      "post": {
        "tags": [
          "Recipe Management"
        ],
        "operationId": "encode_recipe",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EncodeRecipeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Recipe encoded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EncodeRecipeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/recipes/list": {
      "get": {
        "tags": [
          "Recipe Management"
        ],
        "operationId": "list_recipes",
        "responses": {
          "200": {
            "description": "Get recipe list successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListRecipeResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing API key"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/recipes/scan": {
      "post": {
        "tags": [
          "Recipe Management"
        ],
        "operationId": "scan_recipe",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanRecipeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Recipe scanned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanRecipeResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Author": {
        "type": "object",
        "properties": {
          "contact": {
            "type": "string",
            "nullable": true
          },
          "metadata": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "AuthorRequest": {
        "type": "object",
        "properties": {
          "contact": {
            "type": "string",
            "nullable": true
          },
          "metadata": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "ContextManageRequest": {
        "type": "object",
        "description": "Request payload for context management operations",
        "required": [
          "messages",
          "manageAction"
        ],
        "properties": {
          "manageAction": {
            "type": "string",
            "description": "Operation to perform: \"truncation\" or \"summarize\""
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Message"
            },
            "description": "Collection of messages to be managed"
          }
        }
      },
      "ContextManageResponse": {
        "type": "object",
        "description": "Response from context management operations",
        "required": [
          "messages",
          "tokenCounts"
        ],
        "properties": {
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Message"
            },
            "description": "Processed messages after the operation"
          },
          "tokenCounts": {
            "type": "array",
            "items": {
              "type": "integer",
              "minimum": 0
            },
            "description": "Token counts for each processed message"
          }
        }
      },
...
      "CreateRecipeRequest": {
        "type": "object",
        "required": [
          "messages",
          "title",
          "description"
        ],
        "properties": {
          "activities": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "author": {
            "allOf": [
              {
                "$ref": "#/components/schemas/AuthorRequest"
              }
            ],
            "nullable": true
          },
          "description": {
            "type": "string"
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Message"
            }
          },
          "title": {
            "type": "string"
          }
        }
      },
      "CreateRecipeResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "nullable": true
          },
          "recipe": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Recipe"
              }
            ],
            "nullable": true
          }
        }
      },
...
      "DecodeRecipeRequest": {
        "type": "object",
        "required": [
          "deeplink"
        ],
        "properties": {
          "deeplink": {
            "type": "string"
          }
        }
      },
      "DecodeRecipeResponse": {
        "type": "object",
        "required": [
          "recipe"
        ],
        "properties": {
          "recipe": {
            "$ref": "#/components/schemas/Recipe"
          }
        }
      },
      "DeleteRecipeRequest": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "string"
          }
        }
      },
...
      "EncodeRecipeRequest": {
        "type": "object",
        "required": [
          "recipe"
        ],
        "properties": {
          "recipe": {
            "$ref": "#/components/schemas/Recipe"
          }
        }
      },
      "EncodeRecipeResponse": {
        "type": "object",
        "required": [
          "deeplink"
        ],
        "properties": {
          "deeplink": {
            "type": "string"
          }
        }
      },
...
      "ListRecipeResponse": {
        "type": "object",
        "required": [
          "recipe_manifest_responses"
        ],
        "properties": {
          "recipe_manifest_responses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecipeManifestResponse"
            }
          }
        }
      },
...
      "Message": {
        "type": "object",
        "description": "A message to or from an LLM",
        "required": [
          "role",
          "content"
        ],
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MessageContent"
            }
          },
          "created": {
            "type": "integer",
            "format": "int64"
          },
          "id": {
            "type": "string",
            "nullable": true
          },
          "metadata": {
            "$ref": "#/components/schemas/MessageMetadata"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          }
        }
      },
      "MessageContent": {
        "oneOf": [
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/TextContent"
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "text"
                    ]
                  }
                }
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/ImageContent"
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "image"
                    ]
                  }
                }
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/ToolRequest"
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "toolRequest"
                    ]
                  }
                }
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/ToolResponse"
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "toolResponse"
                    ]
                  }
                }
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/ToolConfirmationRequest"
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "toolConfirmationRequest"
                    ]
                  }
                }
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/FrontendToolRequest"
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "frontendToolRequest"
                    ]
                  }
                }
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/ThinkingContent"
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "thinking"
                    ]
                  }
                }
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/RedactedThinkingContent"
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "redactedThinking"
                    ]
                  }
                }
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/ContextLengthExceeded"
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "contextLengthExceeded"
                    ]
                  }
                }
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/SummarizationRequested"
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "summarizationRequested"
                    ]
                  }
                }
              }
            ]
          }
        ],
        "description": "Content passed inside a message, which can be both simple content and tool content",
        "discriminator": {
          "propertyName": "type"
        }
      },
      "MessageMetadata": {
        "type": "object",
        "description": "Metadata for message visibility",
        "properties": {
          "agentVisible": {
            "type": "boolean",
            "description": "Whether the message should be included in the agent's context window"
          },
          "userVisible": {
            "type": "boolean",
            "description": "Whether the message should be visible to the user in the UI"
          }
        }
      },
...
      "Recipe": {
        "type": "object",
        "description": "A Recipe represents a personalized, user-generated agent configuration that defines\nspecific behaviors and capabilities within the Daisy system.\n\n# Fields\n\n## Required Fields\n* `version` - Semantic version of the Recipe file format (defaults to \"1.0.0\")\n* `title` - Short, descriptive name of the Recipe\n* `description` - Detailed description explaining the Recipe's purpose and functionality\n* `Instructions` - Instructions that defines the Recipe's behavior\n\n## Optional Fields\n* `prompt` - the initial prompt to the session to start with\n* `extensions` - List of extension configurations required by the Recipe\n* `context` - Supplementary context information for the Recipe\n* `activities` - Activity labels that appear when loading the Recipe\n* `author` - Information about the Recipe's creator and metadata\n* `parameters` - Additional parameters for the Recipe\n* `response` - Response configuration including JSON schema validation\n* `retry` - Retry configuration for automated validation and recovery\n# Example\n\n\nuse daisy::recipe::Recipe;\n\n// Using the builder pattern\nlet recipe = Recipe::builder()\n.title(\"Example Agent\")\n.description(\"An example Recipe configuration\")\n.instructions(\"Act as a helpful assistant\")\n.build()\n.expect(\"Missing required fields\");\n\n// Or using struct initialization\nlet recipe = Recipe {\nversion: \"1.0.0\".to_string(),\ntitle: \"Example Agent\".to_string(),\ndescription: \"An example Recipe configuration\".to_string(),\ninstructions: Some(\"Act as a helpful assistant\".to_string()),\nprompt: None,\nextensions: None,\ncontext: None,\nactivities: None,\nauthor: None,\nsettings: None,\nparameters: None,\nresponse: None,\nsub_recipes: None,\nretry: None,\n};\n",
        "required": [
          "title",
          "description"
        ],
        "properties": {
          "activities": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "author": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Author"
              }
            ],
            "nullable": true
          },
          "context": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "description": {
            "type": "string"
          },
          "extensions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExtensionConfig"
            },
            "nullable": true
          },
          "instructions": {
            "type": "string",
            "nullable": true
          },
          "parameters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecipeParameter"
            },
            "nullable": true
          },
          "prompt": {
            "type": "string",
            "nullable": true
          },
          "response": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Response"
              }
            ],
            "nullable": true
          },
          "retry": {
            "allOf": [
              {
                "$ref": "#/components/schemas/RetryConfig"
              }
            ],
            "nullable": true
          },
          "settings": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Settings"
              }
            ],
            "nullable": true
          },
          "sub_recipes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SubRecipe"
            },
            "nullable": true
          },
          "title": {
            "type": "string"
          },
          "version": {
            "type": "string"
          }
        }
      },
      "RecipeManifestResponse": {
        "type": "object",
        "required": [
          "name",
          "isGlobal",
          "recipe",
          "lastModified",
          "id"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "isGlobal": {
            "type": "boolean"
          },
          "lastModified": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "recipe": {
            "$ref": "#/components/schemas/Recipe"
          }
        }
      },
...
      "RedactedThinkingContent": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "string"
          }
        }
      },
...
      "Response": {
        "type": "object",
        "properties": {
          "json_schema": {
            "nullable": true
          }
        }
      },
...
      "Role": {
        "oneOf": [
          {
            "type": "string"
          },
          {
            "type": "string"
          }
        ]
      },
...
      "ScanRecipeRequest": {
        "type": "object",
        "required": [
          "recipe"
        ],
        "properties": {
          "recipe": {
            "$ref": "#/components/schemas/Recipe"
          }
        }
      },
      "ScanRecipeResponse": {
        "type": "object",
        "required": [
          "has_security_warnings"
        ],
        "properties": {
          "has_security_warnings": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}